vibeframe-2.0/requirements.txt
# Core dependencies
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.7.0
pandas>=2.0.0

# Web interface
gradio>=4.0.0

# Video processing
moviepy>=2.0.0
ffmpeg-python>=0.2.0
opencv-python>=4.5.0
imageio>=2.22.0

# AI/ML models
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.20.0

# API clients
requests>=2.28.0
aiohttp>=3.8.0

# Data handling
Pillow>=9.0.0,<12.0.0
python-dotenv>=0.19.0

# Testing
pytest>=7.0.0
hypothesis>=6.70.0

# Utilities
tqdm>=4.64.0
pydantic>=2.0.0